0.0.3

+++ Sources/VOISFramework.xcframework/ios-arm64_x86_64-simulator/VOISFramework.framework/Headers/TestRandom.h
@@ -14,5 +14,4 @@
 
 #endif /* TestRandom_h */
 
-//int random_number(int min_num, int max_num);
-double random_number(double min_num, double max_num);
+int random_number(int min_num, int max_num);
